moved the constants to a constants file
